feat(heatmap): add heatmap sub-layer for collection layers
Heatmap layers render the footprints of the products within a collection in a heatmap: altering the color depending on how many footprints intersect at a given pixel. The "styles" paramater alters the colorscale applied, whereas the "dim_range" changes the data range for the colorscale.

Updated documentation for heatmap layers.
Added simple test for heatmap layer.